Allow usage of directives as element attributes
The Chrome inspector (for example) does not handle unrecognized HTML
elements well.

Where one would expect the directive elements to expand and encompass
child elements (children being the svg element), the inspector shows
a height and width of NaN and the elements appear to be collapsed
when inspecting them.

This makes it difficult (if not impossible) to set static heights and
widths on the directive elements.