Add initial unit test and requisite infrastructure
This is the first is what we hope is many commits that include unit
tests with the changes being made.  Currently there is a breakage
for how the getD3Selector method works when the element does not have
its own "id".  We take advantage of this situation to create a concrete
reason to start using unit tests and so this commit presents the needed
groundwork.  We have created an initial test that fails in that case.
The actual fix can then show that this test passes when submitted.

At this point, we document that developer setup required:

  npm install
  grunt bower-install

We add a "test" target to grunt:

  grunt test

We also update a number of the dependencies to resolve mismatches, and
in turn update the stored examples.